+++ CHANGELOG.md
@@ -11,12 +11,14 @@
 * Adding `attributes` to `attributesUpdated` events.
 * Adding `attributes` to `nodeAttributesUpdated` events.
 * Adding `attributes` to `edgeAttributesUpdated` events.
+* Optimizing `*AttributesUpdated` events payload.
 * Optimizing `#.edges(A, B)`.
 * Optimizing `#.forEachEdge(A, B)`.
 * Optimizing `#.mergeEdge`.
 * Fixing `#.edges(A, B)` with directed self loops.
 * Fixing case when `nodeAdded` events would not be fired when using `#.emptyCopy`.
 * Fixing `#.mergeEdgeWithKey` edge cases.
+* Fixing `#.mergeAttributes` event payload.
 * Dropping `before` and `after` from `replace` events metadata.
 * Dropping `value` from `set` events metadata.
 * Correctly emitting `nodeAttributesUpdated` event when node is merged.